Add datastream to schema
Add datastream to schema for Avro.

+++ backend/datahubhel/gateway_utils.py
@@ -88,7 +88,7 @@ def get_kafka_producer() -> confluent_kafka.avro.AvroProducer:
             {"name": "id", "type": "string"},
             {"name": "time", "type": ["null", "long"],
              "logicalType": "timestamp-millis"},
-            {"name": "datastream", "type": "long"},
+            {"name": "datastream", "type": "int"},
             {"name": "value", "type": "string"}
         ]
     }))
